Added LOA processing
When a use sends a message in the leave of absence channel, the bot will attempt to parse a useful date range from the message. If the message does not contain a recognizable and allowable date range (4-12 weeks), the bot will delete the message and alert the user to the cause of the error. Upon receiving a good date range, the bot will create a record in LeaveTimes, or if a record for that officer_id already exists, it will update it. Record deletion is not currently supported.